unprintable

+++ source/assert/h2_timer.cpp
@@ -1,4 +1,4 @@
-h2_inline h2_timer::h2_timer(long long ms_, const char* file_, int line_) : file(file_), line(line_), ms(ms_)
+h2_inline h2_timer::h2_timer(int ms_, const char* file_, int line_) : file(file_), line(line_), ms(ms_)
 {
    start = ::clock();
 }
@@ -6,10 +6,10 @@ h2_inline h2_timer::h2_timer(long long ms_, const char* file_, int line_) : file
 h2_inline h2_timer::~h2_timer()
 {
    h2_assert_g();
-   long long delta = (::clock() - start) * 1000 / CLOCKS_PER_SEC;
+   double delta = (::clock() - start) * 1000.0 / CLOCKS_PER_SEC;
    if (ms < delta) {
       h2_row row = "performance expect < ";
-      row.printf("green", "%lld", ms).printf("", " ms, but actually cost ").printf("red", "%lld", delta).printf("", " ms");
+      row.printf("green", "%d", ms).printf("", " ms, but actually cost ").printf("red", "%d", (int)delta).printf("", " ms");
       h2_fail_g(h2_fail::new_normal(row, file, line));
    }
 }
